// src/app/core/enums/login-steps.enum.ts
export enum LoginSteps {
  CHOOSE_USER = 'chooseUser',
  RESTORE_CHOOSE = 'restoreChoose',
  RESTORE_PHONE = 'restorePhone',
  RESTORE_PHONE_CONFIRM = 'restorePhoneConfirm',
  RESTORE_PHONE_PWD = 'restorePhonePassword',
  RESTORE_EMAIL = 'restoreEmail',
  RESTORE_EMAIL_CODE = 'restoreEmailCode',
  RESTORE_EMAIL_PWD = 'restoreEmailPassword',
  LEGALS = 'legals',
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  EMAIL = 'email',
  PHONE = 'phone',
  RUT_NUMBER = 'rutNumber',
  CLIENT_NUMBER = 'clientNumber',
  PASSWORD = 'password',
  CONFIRM_CODE = 'confirmCode',
}